+++ tests/test_example_7_11.py
@@ -13,7 +13,7 @@ class TestApp(unittest.TestCase):
             index_col=0,
             dtype={"Step": "int32", "Time [days]": "int32"},
         )
-        df_desired.index = df_desired.index.astype(int)
+        # df_desired.index = df_desired.index.astype("int32")
         # vectorize, dense:
         model = create_model(sparse=False)
         model.run(nsteps=27, vectorize=True, threading=True)
@@ -43,7 +43,6 @@ class TestApp(unittest.TestCase):
             drop_nan=True,
             drop_zero=True,
         )
-        df.index = df.index.astype(int)
         pd.testing.assert_frame_equal(df, df_desired)
         np.testing.assert_almost_equal(model.solution.error, 3.450062457943659e-11)
         # symbolic, dense:
@@ -75,7 +74,6 @@ class TestApp(unittest.TestCase):
             drop_nan=True,
             drop_zero=True,
         )
-        df.index = df.index.astype(int)
         pd.testing.assert_frame_equal(df, df_desired)
         np.testing.assert_almost_equal(model.solution.error, 3.450062457943659e-11)
